Remove cast in call to fatal_posix_service_status_errno.

+++ testsuites/psxtests/psxmsgq02/init.c
@@ -50,7 +50,7 @@ void *POSIX_Init(
   attr.mq_msgsize = MSGSIZE;
   puts("Init: mq_open - Workspace not available - ENOMEM");
   mq = mq_open( Get_Longest_Name(), O_CREAT, 0x777, &attr );
-  fatal_posix_service_status_errno((int)mq, ENOMEM, "no workspace available");  
+  fatal_posix_service_status_errno(mq, ENOMEM, "no workspace available");  
 
   puts( "*** END OF POSIX MESSAGE QUEUE 02 TEST ***" );
   rtems_test_exit( 0 );
